examples/apps/sql_agent/pyproject.toml
[project]
name = "sql-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno==1.3.4",
    "altair==5.5.0",
    "annotated-types==0.7.0",
    "anthropic==0.49.0",
    "anyio==4.9.0",
    "attrs==25.3.0",
    "blinker==1.9.0",
    "cachetools==5.5.2",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "distro==1.9.0",
    "docstring-parser==0.16",
    "gitdb==4.0.12",
    "gitpython==3.1.44",
    "google-auth==2.39.0",
    "google-genai==1.11.0",
    "groq==0.22.0",
    "h11==0.14.0",
    "httpcore==1.0.8",
    "httpx==0.28.1",
    "idna==3.10",
    "jinja2==3.1.6",
    "jiter==0.9.0",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2024.10.1",
    "lancedb>=0.22.0",
    "mapepire-python>=0.2.0",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mcp[cli]>=1.6.0",
    "mdurl==0.1.2",
    "narwhals==1.35.0",
    "nest-asyncio==1.6.0",
    "numpy==2.2.5",
    "ollama>=0.4.8",
    "openai==1.75.0",
    "packaging==24.2",
    "pandas==2.2.3",
    "pgvector==0.4.0",
    "pillow==11.2.1",
    "protobuf==5.29.4",
    "psycopg==3.2.6",
    "psycopg-binary==3.2.6",
    "pyarrow==19.0.1",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.11.3",
    "pydantic-core==2.33.1",
    "pydantic-settings==2.9.1",
    "pydeck==0.9.1",
    "pygments==2.19.1",
    "pylance>=0.26.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "referencing==0.36.2",
    "requests==2.32.3",
    "rich==14.0.0",
    "rpds-py==0.24.0",
    "rsa==4.9.1",
    "shellingham==1.5.4",
    "simplejson==3.20.1",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.40",
    "streamlit==1.44.1",
    "tantivy>=0.22.2",
    "tenacity==9.1.2",
    "toml==0.10.2",
    "tomli==2.2.1",
    "tornado==6.4.2",
    "tqdm==4.67.1",
    "typer==0.15.2",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.0",
    "tzdata==2025.2",
    "urllib3==2.4.0",
    "websockets==15.0.1",
]
